+++ src/app/main.cpp
@@ -62,6 +62,7 @@
 #include "effects/vst/vsteffectsmodule.h"
 #endif
 #include "effects/nyquist/nyquisteffectsmodule.h"
+#include "importexport/import/importermodule.h"
 
 #include "au3wrap/au3wrapmodule.h"
 
@@ -155,6 +156,7 @@ int main(int argc, char** argv)
     app.addModule(new au::effects::VstEffectsModule());
 #endif
     app.addModule(new au::effects::NyquistEffectsModule());
+    app.addModule(new au::importexport::ImporterModule());
 
 #if (defined (_MSCVER) || defined (_MSC_VER))
     // On MSVC under Windows, we need to manually retrieve the command-line arguments and convert them from UTF-16 to UTF-8.
