+++ src/importexport/import/CMakeLists.txt
@@ -1,0 +1,27 @@
+#
+# Audacity: A Digital Audio Editor
+#
+
+declare_module(importer)
+
+set(MODULE_SRC
+    # public
+    ${CMAKE_CURRENT_LIST_DIR}/importermodule.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/importermodule.h
+    ${CMAKE_CURRENT_LIST_DIR}/iimporter.h
+
+    #internals
+    ${CMAKE_CURRENT_LIST_DIR}/internal/au3/au3importer.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/au3/au3importer.h
+)
+
+# AU3
+include(${CMAKE_CURRENT_LIST_DIR}/../../au3wrap/au3defs.cmake)
+
+set(MODULE_INCLUDE ${AU3_INCLUDE})
+set(MODULE_DEF ${AU3_DEF})
+set(MODULE_LINK au3wrap)
+
+set(MODULE_USE_UNITY OFF)
+
+setup_module()
