+++ src/importexport/import/iimporter.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "io/path.h"
+#include "modularity/imoduleinterface.h"
+
+namespace au::importexport {
+class IImporter : MODULE_EXPORT_INTERFACE
+{
+    INTERFACE_ID(IImporter)
+
+public:
+    virtual ~IImporter() = default;
+
+    virtual bool import(const muse::io::path_t& filePath) = 0;
+};
+}
